dashboards/requirements.txt
altair==5.5.0 ; python_version >= "3.11" and python_version < "4.0"
attrs==25.3.0 ; python_version >= "3.11" and python_version < "4.0"
blinker==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
cachetools==5.5.2 ; python_version >= "3.11" and python_version < "4.0"
certifi==2025.7.14 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.4.2 ; python_version >= "3.11" and python_version < "4.0"
click==8.2.1 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
db-dtypes==1.4.3 ; python_version >= "3.11" and python_version < "4.0"
gitdb==4.0.12 ; python_version >= "3.11" and python_version < "4.0"
gitpython==3.1.44 ; python_version >= "3.11" and python_version < "4.0"
google-api-core==2.25.1 ; python_version >= "3.11" and python_version < "4.0"
google-auth-oauthlib==1.2.2 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.40.3 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-bigquery==3.34.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-core==2.4.3 ; python_version >= "3.11" and python_version < "4.0"
google-crc32c==1.7.1 ; python_version >= "3.11" and python_version < "4.0"
google-resumable-media==2.7.2 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos==1.70.0 ; python_version >= "3.11" and python_version < "4.0"
grpcio-status==1.73.1 ; python_version >= "3.11" and python_version < "4.0"
grpcio==1.73.1 ; python_version >= "3.11" and python_version < "4.0"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.11" and python_version < "4.0"
jsonschema-specifications==2025.4.1 ; python_version >= "3.11" and python_version < "4.0"
jsonschema==4.24.0 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.11" and python_version < "4.0"
narwhals==1.47.0 ; python_version >= "3.11" and python_version < "4.0"
numpy==2.3.1 ; python_version >= "3.11" and python_version < "4.0"
oauthlib==3.3.1 ; python_version >= "3.11" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.11" and python_version < "4.0"
pandas-gbq==0.29.2 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.3.1 ; python_version >= "3.11" and python_version < "4.0"
pillow==11.3.0 ; python_version >= "3.11" and python_version < "4.0"
proto-plus==1.26.1 ; python_version >= "3.11" and python_version < "4.0"
protobuf==6.31.1 ; python_version >= "3.11" and python_version < "4.0"
pyarrow==20.0.0 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.4.2 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.11" and python_version < "4.0"
pydata-google-auth==1.9.1 ; python_version >= "3.11" and python_version < "4.0"
pydeck==0.9.1 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
pytz==2025.2 ; python_version >= "3.11" and python_version < "4.0"
referencing==0.36.2 ; python_version >= "3.11" and python_version < "4.0"
requests-oauthlib==2.0.0 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.4 ; python_version >= "3.11" and python_version < "4.0"
rpds-py==0.26.0 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9.1 ; python_version >= "3.11" and python_version < "4.0"
setuptools==80.9.0 ; python_version >= "3.11" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.11" and python_version < "4.0"
smmap==5.0.2 ; python_version >= "3.11" and python_version < "4.0"
streamlit==1.47.0 ; python_version >= "3.11" and python_version < "4.0"
tenacity==9.1.2 ; python_version >= "3.11" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.11" and python_version < "4.0"
tornado==6.5.1 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.14.1 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2025.2 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.5.0 ; python_version >= "3.11" and python_version < "4.0"
watchdog==6.0.0 ; python_version >= "3.11" and python_version < "4.0" and platform_system != "Darwin"
